BCK-22-Dodavanje permission i role repo controller i service
