+++ user-service/src/main/java/rs/edu/raf/userservice/domains/dto/employee/EmployeeCreateDto.java
@@ -1,5 +1,8 @@
-package rs.edu.raf.userservice.domains.dto;
+package rs.edu.raf.userservice.domains.dto.employee;
 
+//import jakarta.validation.constraints.Email;
+//import jakarta.validation.constraints.NotBlank;
+//import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
@@ -13,7 +16,8 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 @RequiredArgsConstructor
-public class EmployeeUpdateDto {
+public class EmployeeCreateDto {
+
     @NotBlank
     private String firstName;
     @NotBlank
@@ -33,14 +37,17 @@ public class EmployeeUpdateDto {
     private String address;
 //    @NotBlank
 //    private String password;
+//    @NotBlank
+//    private String saltPassword;
     @NotBlank
     private String department;
     @NotBlank
     private String position;
     @NotBlank
     private String phoneNumber;
-    @NotNull
-    private Boolean isActive;
+
+//    private Boolean isActive;
+
     @NotNull
     private Role roles;
 }
