+++ user-service/src/main/java/rs/edu/raf/userservice/domains/dto/user/UpdateUserDto.java
@@ -1,4 +1,4 @@
-package rs.edu.raf.userservice.domains.dto;
+package rs.edu.raf.userservice.domains.dto.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,12 +7,13 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class CreateUserDto {
+public class UpdateUserDto {
     private String firstName;
     private String lastName;
-    private String jmbg;
-    private String dateOfBirth;
+    private Long dateOfBirth;
     private String gender;
     private String phoneNumber;
+    private String address;
     private String email;
+    private Boolean isActive;
 }
