+++ user-service/src/main/java/rs/edu/raf/userservice/domains/model/Role.java
@@ -19,7 +19,6 @@ public class Role implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @JsonIgnore
     private Long roleId;
 
     @Enumerated(EnumType.STRING)
