+++ user-service/src/main/java/rs/edu/raf/userservice/repositories/EmployeeRepository.java
@@ -2,10 +2,8 @@ package rs.edu.raf.userservice.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import rs.edu.raf.userservice.domains.dto.EmployeeCreateDto;
 import rs.edu.raf.userservice.domains.model.Employee;
 
-import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
