+++ user-service/src/test/java/rs/edu/raf/userservice/EmployeeServiceUnitTests.java
@@ -1,18 +1,11 @@
 package rs.edu.raf.userservice;
 
-import io.cucumber.java.an.E;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import rs.edu.raf.userservice.domains.dto.EmployeeDto;
-import rs.edu.raf.userservice.domains.model.Employee;
 import rs.edu.raf.userservice.repositories.EmployeeRepository;
 import rs.edu.raf.userservice.services.EmployeeService;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
 public class EmployeeServiceUnitTests {
